Refactor chats toward shared code
